Binary Masks Have Been Made and Data Has Been Broken Up
I've made the binary masks for the mri images that have manual contours. Also, I wanted to upload the data with it in case I lose it on my computer. The data is not in the original format that it was in when it was first retrieved from the website. I wanted to make it easier for a script to associate a patient's contour with their mri image, so I put the contours in the same directory as the patient it belongs to.

Next steps are to make and train a model that follows the u-net architecture.