Model trained and a minimum loss was reached
A new dataset for validation was obtained and tuned into a workable dataset. I also made modifications to the notebook that makes binary masks. It now fills in the manual contours with 1s instead of just creating a boundary. This caused problems for training the model. I believe this has been fixed and will use the validation set to see if the model learned.