Model is done. Will continue hyperparameter tuning. Next Thing is to see how other models from the Monai zoo do on the data
My take on the U-net model is done and seems to be working pretty well. The model does seem to predict a larger area than the target masks (by typically a factor of 2), however, I believe this is very likely due to the high weight given to the minority 1s class. The model predicts the right area as one can see in the Results.zip file.

I am going to tune the hyperparameters to see if the area gets smaller. I will also use a couple models from the Monai Zoo that were trained on cardiac data to see if they do any better (That will be in another coming notebook titles otherModels.ipynb).