Final Push
The notebook which contains a majority of the project code has been cleaned up and fixed. I measured the distance of the area predicted by the mask wrong before, now the function works as intended and measures real diameters of the left ventricle. I've also uploaded the last time sheet.